+++ src/routes/(MainApp)/[userEmail]/[subjectID]/[workspaceID]/+page.server.js
@@ -1,14 +1,9 @@
 //@ts-ignore
 import prisma from '$lib/db';
-import { error, invalid, redirect } from '@sveltejs/kit';
-import { get } from 'svelte/store';
-import { global_PASS, global_USERID } from '$lib/stores/global.store';
-import bcryptjs from 'bcryptjs';
+import { error, invalid, redirect } from '@sveltejs/kit'
 
 /** @type {import('./$types').PageServerLoad} */
 export async function load({ params }) {
-	if (!get(global_PASS)) throw redirect(303, '/Signin');
-
 	const user = await prisma.users.findFirst({
 		where: {
 			email: {
@@ -16,8 +11,7 @@ export async function load({ params }) {
 			}
 		}
 	});
-	if (!user) throw error(404, 'Account not found');
-	if (!bcryptjs.compareSync(get(global_PASS), user.password)) throw error(401, 'Unauthorized access');
+	if (!user) throw error(404, 'Account not found')
 
 	const subject = await prisma.subjects.findFirst({
 		where: {
@@ -185,7 +179,6 @@ export const actions = {
 				}
 			}
 		});
-
 		if (!user) throw error(404, 'Account not found');
 
 		const user2 = await prisma.users.update({
@@ -198,9 +191,8 @@ export const actions = {
 				}
 			}
 		});
-
 		if (!user2) throw redirect(301, 'my-profile');
-		global_PASS.set('');
+		
 		throw redirect(301, '/Signin');
 	},
 	updateWorkspace: async ({ request, params }) => {
@@ -242,7 +234,6 @@ export const actions = {
         id: true
       }
     })
-
     if(!upatedUser) return invalid(500, {message: 'Failed to update favorites but is added to favorites, please reload', reason: 'databaseError'})
 
 		const updatedWorkspace = await prisma.workspaces.update({
@@ -257,7 +248,6 @@ export const actions = {
 				id: true
 			}
 		});
-
 		if (!updatedWorkspace)
 			return invalid(500, {
 				message: 'Database failure to update workspace',
@@ -273,10 +263,12 @@ export const actions = {
 		const status = data.get('status')?.toString();
 		const dueDateTime = data.get('dueDateTime')?.toString();
 		const members = data.get('members')?.toString();
+		const creator = data.get('creator')?.toString()
 
 		const newTask = await prisma.tasks.create({
 			data: {
-				createdBy: get(global_USERID),
+				//@ts-ignore
+				createdBy: creator,
 				createdOn: new Date(),
 				//@ts-ignore
 				description,
@@ -292,14 +284,15 @@ export const actions = {
 				//@ts-ignore
 				members: members?.length > 0 ? members?.split(',') : [],
 				subtasks: [],
-				viewers: [get(global_USERID)],
-				subscribers: [get(global_USERID)]
+				//@ts-ignore
+				viewers: [creator],
+				//@ts-ignore
+				subscribers: [creator]
 			},
 			select: {
 				id: true
 			}
 		});
-
 		if (!newTask.id)
 			return invalid(403, {
 				message: 'Database failure to create a task',
@@ -319,7 +312,6 @@ export const actions = {
 				id: true
 			}
 		});
-
 		if (!updatedBoard.id)
 			return invalid(500, {
 				message: 'Task created but failed to be syncronized in boards, try again',
@@ -331,6 +323,7 @@ export const actions = {
 		const name = data.get('name')?.toString();
 		const description = data.get('description')?.toString();
 		const status = data.get('status')?.toString();
+		const oldStatus = data.get('oldStatus')?.toString()
 		//@ts-ignore
 		const level = parseInt(data.get('level')?.toString());
 		const dueDateTime = data.get('dueDateTime')?.toString();
@@ -349,28 +342,55 @@ export const actions = {
 				status: status
 			},
 			select: {
-				id: true
+				id: true,
+				status: true
 			}
 		});
+		if (!updatedTask.id) return invalid(500, { message: 'Update failure on the task', reason: 'databaseError' });
 
-		if (!updatedTask.id)
-			return invalid(500, { message: 'Update failure on the task', reason: 'databaseError' });
-
-		const updatedBoard = await prisma.boards.update({
-			where: {
-				id: status
-			},
-			data: {
-				tasks: {
-					push: updatedTask.id
+		if (oldStatus !== status) {
+			let toUpdateBoard1 = await prisma.boards.findFirst({
+				where: {
+					//@ts-ignore
+					id: oldStatus
+				},
+				select: {
+					id: true,
+					tasks: true
 				}
-			},
-			select: {
-				id: true
-			}
-		});
+			})
+			if(!toUpdateBoard1) throw error(404, 'Board to update not found')
+			
+			toUpdateBoard1.tasks = toUpdateBoard1.tasks.filter(id => id !== taskID)
 
-		if(!updatedBoard) return invalid(500, {message: 'Updated board data failed to fetch please reload', reason: 'databaseError'})
+			const updatedBoard1 = await prisma.boards.update({
+				where: {
+					id: toUpdateBoard1.id
+				},
+				data: {
+					tasks: toUpdateBoard1.tasks
+				},
+				select: {
+					id: true
+				}
+			})
+			if(!updatedBoard1) throw error(404, 'Failed to fetch updated board data')
+
+			const updatedBoard = await prisma.boards.update({
+				where: {
+					id: status
+				},
+				data: {
+					tasks: {
+						push: updatedTask.id
+					}
+				},
+				select: {
+					id: true
+				}
+			});
+			if(!updatedBoard) return invalid(500, {message: 'Updated board data failed to fetch please reload', reason: 'databaseError'})
+		}
 	},
 	deleteTask: async ({ request }) => {
 		const data = await request.formData();
@@ -387,7 +407,6 @@ export const actions = {
 				status: true
 			}
 		});
-
 		if (!deletedTask.id)
 			return invalid(500, {
 				message: "Can't delete or remove task pleast try again later",
@@ -405,8 +424,7 @@ export const actions = {
 				id: true
 			}
 		});
-
-		if (!toUpdateBoard?.id)
+		if (!toUpdateBoard)
 			return invalid(404, {
 				message: 'Board that a task is in cannot be found, please try again',
 				reason: 'databaseError'
@@ -417,9 +435,10 @@ export const actions = {
 				id: toUpdateBoard.id
 			},
 			data: {
-				tasks: toUpdateBoard?.tasks.filter((id) => id !== deletedTask.id)
+				tasks: toUpdateBoard.tasks.filter((id) => id !== deletedTask.id)
 			}
 		});
+		if(!updatedBoard) throw error(404, 'Failed to fetch data of board, please try again later')
 
 		const deletedSUBTasks = prisma.tasks.deleteMany({
 			where: {
